chce_postgress_backup fix

+++ scripts/chce_postgres_backup.sh
@@ -15,7 +15,7 @@ DB_USER="postgres"
 DB_PASSWORD="xxx"
 
 #Numbers of days you want to keep copy of your databases
-NUMBER_OF_DAYS=180
+NUMBER_OF_DAYS=30
 echo "Dumping database to ${BACKUP_DIR}${DB_NAME}_${BACKUP_DATE}.sql" 
 
 if [ ! -d "${BACKUP_DIR}" ]; then
@@ -24,6 +24,14 @@ fi
    
 PGPASSWORD="${DB_PASSWORD}" pg_dump -h "${DB_HOST}" -p "${DB_PORT}" -U "${DB_USER}" -d "${DB_NAME}" -w --format=custom | xz > "${BACKUP_DIR}${DB_NAME}_${BACKUP_DATE}.xz"
 
-echo "Dumping database finnish" 
+# Attempt to create the backup
+if PGPASSWORD="${DB_PASSWORD}" pg_dump -h "${DB_HOST}" -p "${DB_PORT}" -U "${DB_USER}" -d "${DB_NAME}" -w --format=custom | xz > "${BACKUP_DIR}${DB_NAME}_${BACKUP_DATE}.xz"; then
+    echo "Dumping database finished successfully"
+    
+    # Delete old backups, but only if a current backup exists
+    find "${BACKUP_DIR}" -type f -prune -mtime +"${NUMBER_OF_DAYS}" -exec rm -f {} \;
+    echo "Old backups deleted"
+else
+    echo "Error: Dumping database failed"
+fi
 
-find "${BACKUP_DIR}" -type f -prune -mtime +"${NUMBER_OF_DAYS}" -exec rm -f {} \;
